docs: update Azure Key Vault entry
The generated password from `az` does no longer appear to be a UUID,
but returns a random generated string instead.

Given there is also a UUID attached to Secret object that holds this
value, it is better to be explict to guide users.

In addition, the tenant is always a UUID (and not _any_ ID).

Signed-off-by: Hidde Beydals <hello@hidde.co>